OMAP4430: Fix errata i723 (DPLL ABE Warm Reset Issue)
Linux hangs during boot after warm reset. The M (multiplier) and N (divider)
values for the DPLL_ABE are programmed inside the x-loader. After post warm
reset the DPLL_ABE is programmed to go in MN bypass state. The M and N values
for the DPLL_ABE remain at 0x0 inspite of the programmation done by SW.
The PLL is now programmed to go in Locked state. The PLL is not able to lock
even if M and N are programmed in PRCM register causing a HANG.

Work around: Re-programmation of the M and N values of DPLL_ABE needs to be
done when DPLL_EN bit is 0x7. Re-programation will not work if DPLL_EN bit
is 0x4 and PLL is in stop/idle-bypass state.

Therefore M and N are reprogrammed for the DPLL_ABE after going to Lock state
(in this case M and N values reset becomes inactive).

Change-Id: I354fd18b8923984d9b03d5df3369486429868ab1
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>