OMAP4: Correct init values for EMIF registers.
EMIF_SDRAM_CONFIG register: The values of "REG_DDR2_DDQS" bit only applies to
DDR2, not to LPDDR2. This value is reserved/not required for OMAP4 as DDR2 is not
supported. However as per recommendation from HW designer, this bit is set to '1' for
both 4430 and 4460 chipsets.

EMIF_DDR_PHY_CTRL_1 register: The value of "DLL_MASTER_SENSITIVITY" bit is set to
0x0. The original recommendation was value 0x2 as per the TRM; however most of the
silicon validation has been performed with value 0x0. Hence as recommended from HW
designer the bit is set to zero for both 4430 and 4460 chipsets.

The structure "ddr_regs_elpida2G_400_mhz_2cs_omap4460" has been removed from the
code and a single structure "ddr_regs_elpida2G_400_mhz_2cs" is now maintained for both
4430 and 4460.

Change-Id: I99f9531a4f864acb0202229c535db2e545be8fae
Signed-off-by: Priyaranjan Das <x0159917@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>