OMAP4: PM: voltage: split voltage scale per CPU
When the original code was written, it was meant to support a single
CPU - OMAP4430. With the addition of OMAP4460 and OMAP4470, code
is a nightmarish maze of if-else-goto conditions making the logic
un-maintainable.

Split the sequence into per OMAP SoC scale function where appropriate
handling can be done per SoC-PMIC combination.

NOTE: This does not make life easier for platforms with custom PMICs
or non-standard combinations, but that is something to be either
handled in forks, this patch currently supports just the TI
recommended PMIC combination per SoC.

Change-Id: I2c49bdadc0da3df433c688033d38dbd699b2edd3
Reported-by: Isabelle Gros <i-gros@ti.com>
Reported-by: Jerome Angeloni <j-angeloni@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>