OMAP4: EMIF: fix self refresh rate after warm reset
Errata ID: i727

Description:
When a warm reset is applied on the system, the OMAP processor
restarts with another OPP and so frequency is not the same.
Due to this frequency change, the refresh rate programmed is not
the accurate value and could result in an unexpected behavior
on the memory side.

Workaround:
The workaround is to force self refresh immediately when coming
back from the warm reset.
Set EMIF_PWR_MGMT_CTRL[10:8] REG_LP_MODE to 0x2
Set EMIF_PWR_MGMT_CTRL[7:4] REG_SR_TIM to 0x0
Do a dummy read (loads automatically new value of sr_tim)

Change-Id: Ib117ee7a8fede11e5b8478ca7bdf966b78ff464d
Signed-off-by: Kevin McCray <kevin.mccray@ti.com>